third_party/blink/web_tests/external/wpt/speculation-rules/prefetch/redirect-to-prefetch-url.https_origin=cross-site-initial-expected.txt

This is a testharness.js-based test.
[FAIL] redirect-received-before-navigation-start
  assert_equals: Prefetched response should not be used by navigation. expected "" but got "prefetch"
[FAIL] redirect-received-after-navigation-start
  assert_equals: Prefetched response should not be used by navigation. expected "" but got "prefetch"
Harness: the test ran to completion.
